update the make_sym_link_files.py code

+++ general/raw_data/make_sym_link_files.py
@@ -11,14 +11,15 @@ Created on 2021-06-14
 """
 import os
 import glob
+from tqdm import tqdm
 
 # =============================================================================
 # Define variables
 # =============================================================================
 # define raw directory
-RAW_DIR = '/data/spirou/data/common/raw/'
+RAW_DIR = '/spirou/cfht_nights/common/raw'
 # define out directory
-OUT_DIR = '/data/spirou/data/common/rawfilesym/'
+OUT_DIR = '/spirou/drs-data/common/raw'
 
 
 # =============================================================================
@@ -28,18 +29,23 @@ if __name__ == "__main__":
     # get all raw files
     files = glob.glob(os.path.join(RAW_DIR, '*', '*.fits'))
     # loop around files and make links
-    for it, filename in enumerate(files):
+    for it, filename in tqdm(enumerate(files)):
         # get basename
         basename = os.path.basename(filename)
         # get directory
         obs_dir = os.path.basename(os.path.dirname(filename))
         # create outfile
         outfile = os.path.join(OUT_DIR, obs_dir, basename)
-        # print progress
-        margs = [it + 1, len(files), outfile]
-        print('[{0}/{1}] {2}'.format(*margs))
-        # make a symlink
-        os.symlink(filename, outfile)
+        # do not recreate links
+        if not os.path.exists(outfile):
+            # create dir if it doesn't exist
+            if not os.path.exists(os.path.dirname(outfile)):
+                os.mkdir(os.path.dirname(outfile))
+            # print progress
+            margs = [it + 1, len(files), outfile]
+            print('[{0}/{1}] {2}'.format(*margs))
+            # make a symlink
+            os.symlink(filename, outfile)
 
 
 # =============================================================================
